### 2024-09-27 #7295 @editorialbot
[PRE REVIEW]: pyMechT: A Python package for mechanics of soft tissues
**Submitting author:** @ankushaggarwal (<a href="http://orcid.org/0000-0002-1755-8807">Ankush Aggarwal</a>)
**Repository:** https://github.com/ankushaggarwal/pymecht
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.2
