### 2024-09-27 #7295 @Kevin-Mattheus-Moerman
I'll assign myself as editor as soft tissue biomechanics is part of my expertise, and other editors in this domain are currently busy handling other submissions. So I will be the editor responsible for finding reviewers here and you can address me if you have any questions.
