### 2024-11-26 #7490 @jpsferreira
@ankushaggarwal

- I noticed that `torch` is listed as a dependency in the project.toml file, but I could not find any references to it in the source code. Could the authors clarify its role in the project or confirm if it is an unused dependency?
- The authors claim that the repository can simulate mechanical experiments significantly faster than conventional FEM software. However, I could not find a direct comparison of computational time between the two approaches. If such comparisons have been performed in previous works, it would be helpful to reference or summarize them in the manuscript for context.
- The repository lacks community guidelines, such as instructions for user support, contribution procedures, and bug and issue reporting. Adding these guidelines would enhance the project's usability and community engagement.
- The paper adheres to the journal’s structural requirements and explains how the package handles conservative hyperelastic materials. While future extensions to inelastic materials are mentioned, the paper does not outline potential schemes for addressing symbolic differentiation with history variables, which is a complex and non-trivial task. Including such details would strengthen the scope and future impact of the work.
- The manuscript mentions plans to include ex-vivo setups that generate heterogeneous deformations in specimens. Could the authors elaborate on how these cases will be addressed without using FEM? Clarifying this point would provide valuable insights into the methodology’s applicability to more complex setups.
- In the reference list, there is a paper under review authored by the first author. It is unclear if citing a work that is still under review aligns with the journal’s guidelines. I recommend verifying this with the editorial policies and, if necessary, replacing or rephrasing the reference.
