### 2024-11-30 #7490 @ankushaggarwal
@jpsferreira  Thank you for the review and suggestions.

- SobolEngine from `torch` is used in [RandomParameters class](https://github.com/ankushaggarwal/pymecht/blob/a81d59f02045eac3067e380f79b7b4f30241eb15/pymecht/RandomParameters.py#L5). Admittedly, this is only a small functionality out of the whole `torch` package. In the future, it might be better to replace it with another less extensive package (any suggestions would be welcome).
- The "faster" feature is when accounting for setting up the problem (i.e., geometry and mesh creation, etc. in FEA). The computational time on its own may not be much faster, depending on the mesh density. Thus, I am not sure if we can directly compare the computational time. Nevertheless, if there is a particular instance in the documentation/paper that is misleading, we would be happy to modify it to make it clearer.
- Good point. I will work on adding the community guidelines.
- For the future extensions,
  - For inelastic models, it would be best to start with pre-defined models that can implemented numerically along with a time integration scheme, before moving to a general model with symbolic differentiation. We would be happy to add this clarification to the manuscript.
  - By heterogeneous deformation, are you referring to the microindentation? If so, we were only thinking of analytical (closed-form) models such as Hertz model.
 - I will check the editorial policies and also check with the editor for an estimate of when we can expect the DOI, as the cited chapter is under production.
