### 2024-12-19 bf9670d
Updated JOSS paper: updated the chapter reference to in press, added a few more references, and clarified the faster nature and future plans for pymecht
