+++ joss-paper/paper.md
@@ -113,7 +113,7 @@ Detailed documentation is hosted on [`readthedocs`](https://pymecht.readthedocs.
 # Advantages over finite element simulation
 In principle, the problems that can be solved using `pyMechT` can also be solved using any finite element simulation software. However, `pyMechT` offers the following advantages:
 
-- Geometry and mesh creation would be required for a finite element simulation, which usually takes some time. However, the pre-defined geometrical features in `pyMechT` means that one only needs to choose the right class and parameters. In addition, no meshing is required. This means that setting up the problem is much faster.
+- Geometry and mesh creation would be required for a finite element simulation, which usually takes some time. However, the pre-defined geometrical features in `pyMechT` means that one only needs to choose the right class and parameters. In addition, no meshing is required. This means that setting up the problem (i.e., defining the geometry/mesh and loads) is much faster in `pyMechT`. Once the model has been setup, the computational time required to solve it is comparable, depending on the finite element mesh density. 
 
 - Enforcing incompressibility in a finite element simulation can be numerically challenging, necessitating approaches such as Lagrange multiplier with a three-field formulation. Instead, in `pyMechT`, the incompressibility is analytically enforced *exactly*, thus making the results more robust.
 
@@ -122,9 +122,9 @@ In principle, the problems that can be solved using `pyMechT` can also be solved
 - The reference zero-stress state of biological tissues can be unknown or ambiguous. Moreover, the biological tissues are heterogeneous, with multiple layers each of varying properties. These aspects are non-trivial to incorporate in a finite element simulation, due to the need for recreating the geometry and/or incompatability of the initial state. However, it is straightforward to simulate these in `pyMechT`.
 
 # Uses in literature
-`pyMechT` has been used for Bayesian model selection based on extensive planar biaxial extension data [@AGGARWAL2023105657]. This work required rapid simulation of varied constitutive models, which was facilitated by `pyMechT`. Similarly, the Bayesian inference via Markov Chain Monte Carlo in `pyMechT` was used to infer the distribution of aortic biomechanical and geometrical properties based on in-vivo measurements (as likelihood) and ex-vivo biaxial extension data (as prior distribution) [@AGGARWAL2024]. Moreover, data-driven model developed in @AGGARWAL2023115812 has been used in `pyMechT` via the `splineI1` and `splineI1I4` material models. 
+`pyMechT` has been used for Bayesian model selection based on extensive planar biaxial extension data [@AGGARWAL2023105657]. This work required rapid simulation of varied constitutive models, which was facilitated by `pyMechT`. Similarly, the Bayesian inference via Markov Chain Monte Carlo in `pyMechT` was used to infer the distribution of aortic biomechanical and geometrical properties based on in-vivo measurements (as likelihood) and ex-vivo biaxial extension data (as prior distribution) [@Aggarwal2025]. Moreover, data-driven model developed in @AGGARWAL2023115812 has been used in `pyMechT` via the `splineI1` and `splineI1I4` material models. 
 
 # Conclusion and future plans
-`pyMechT` fills an important gap and allows soft tissue biomechanics researchers to model ex-vivo testing setups in a fast, robust, and flexible manner. The package is numerically efficient and extensively documented. It has facilitated several publications, and we believe that it can benefit the wider community. In the future, we plan to extend the capabilities of the package to include more material models, such as inelastic (viscoelastic, plastic, damage, growth & remodeling), and other ex-vivo setups (such as microindentation). Lastly, the package could be coupled with others to allow multi-physics simulations, such as for hemodynamics and biochemical regulation.
[…]
