+++ joss-paper/paper.bib
@@ -20,11 +20,52 @@ doi = {10.1016/j.jmbbm.2023.105657},
 author={Aggarwal, Ankush and Hudson, Luke T and Laurence, Devin W and Lee, Chung-Hao and Pant, Sanjay}
 }
 
-@bookchapter{AGGARWAL2024,
-title = {Heterogeneity and multi-scale modeling in vascular biomechanics},
-booktitle = {Integration and Bridging of Multiscale Bioengineering Designs and Tissue Biomechanics},
-publisher = {Springer},
-year = {under review},
-author={Aggarwal, Ankush and Coccarelli, Alberto and McGinty, Sean},
-editor={Liao, Jun and Wong, Joyce Y}
+@bookchapter{Aggarwal2025,
+  author    = {Aggarwal, Ankush and Coccarelli, Alberto and McGinty, Sean},
+  title     = {Heterogeneity and multi-scale modeling in vascular biomechanics},
+  booktitle = {Integration and Bridging of Multiscale Bioengineering Designs and Tissue Biomechanics},
+  editor    = {Jun Liao and Joyce Y. Wong},
+  publisher = {Springer Nature Switzerland},
+  year      = {2025},
+  %pages     = {xx--yy},  
+  isbn      = {978-3-031-81742-7},
+  note      = {In press},
 }
+
+@article{coccarelli2021framework,
+  title={A framework for incorporating 3D hyperelastic vascular wall models in 1D blood flow simulations},
+  author={Coccarelli, Alberto and Carson, Jason M and Aggarwal, Ankush and Pant, Sanjay},
+  journal={Biomechanics and Modeling in Mechanobiology},
+  volume={20},
+  number={4},
+  pages={1231--1249},
+  year={2021},
+  publisher={Springer},
+  doi = {10.1007/s10237-021-01437-5}
+}
+
+@software{alberto_coccarelli_2021_4522152,
+  author       = {Alberto Coccarelli and
+                  Jason M Carson and
+                  Ankush Aggarwal and
+                  Sanjay Pant},
+  title        = {1D-Hyperelastic-Haemodynamics: Version1},
+  month        = feb,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {V1},
+  doi          = {10.5281/zenodo.4522152},
+}
+
+@article{coccarelli2024new,
+  title={A new model for evaluating pressure-induced vascular tone in small cerebral arteries},
+  author={Coccarelli, Alberto and Pant, Sanjay and Polydoros, Ioannis and Harraz, Osama F},
+  journal={Biomechanics and Modeling in Mechanobiology},
+  volume={23},
+  number={1},
+  pages={271--286},
+  year={2024},
+  publisher={Springer},
+  doi={10.1007/s10237-023-01774-7}
+}
+
