### 2024-12-19 868d314
Added CONTRIBUTING.md and referred to it in README.md
