### 2024-12-19 14dfc8b
Small citation format fix
