+++ joss-paper/paper.md
@@ -125,6 +125,6 @@ In principle, the problems that can be solved using `pyMechT` can also be solved
 `pyMechT` has been used for Bayesian model selection based on extensive planar biaxial extension data [@AGGARWAL2023105657]. This work required rapid simulation of varied constitutive models, which was facilitated by `pyMechT`. Similarly, the Bayesian inference via Markov Chain Monte Carlo in `pyMechT` was used to infer the distribution of aortic biomechanical and geometrical properties based on in-vivo measurements (as likelihood) and ex-vivo biaxial extension data (as prior distribution) [@Aggarwal2025]. Moreover, data-driven model developed in @AGGARWAL2023115812 has been used in `pyMechT` via the `splineI1` and `splineI1I4` material models. 
 
 # Conclusion and future plans
-`pyMechT` fills an important gap and allows soft tissue biomechanics researchers to model ex-vivo testing setups in a fast, robust, and flexible manner. The package is numerically efficient and extensively documented. It has facilitated several publications, and we believe that it can benefit the wider community. In the future, we plan to extend the capabilities of the package to include more material models, such as inelastic (viscoelastic, plastic, damage, growth & remodeling) pre-defined formulations, and other ex-vivo setups (such as microindentation using Hertz contact model). Lastly, the package could be coupled with others to allow multi-physics simulations, such as for hemodynamics [@coccarelli2021framework, @alberto_coccarelli_2021_4522152] and biochemical regulation [@coccarelli2024new].
+`pyMechT` fills an important gap and allows soft tissue biomechanics researchers to model ex-vivo testing setups in a fast, robust, and flexible manner. The package is numerically efficient and extensively documented. It has facilitated several publications, and we believe that it can benefit the wider community. In the future, we plan to extend the capabilities of the package to include more material models, such as inelastic (viscoelastic, plastic, damage, growth & remodeling) pre-defined formulations, and other ex-vivo setups (such as microindentation using Hertz contact model). Lastly, the package could be coupled with others to allow multi-physics simulations, such as for hemodynamics [@coccarelli2021framework; @alberto_coccarelli_2021_4522152] and biochemical regulation [@coccarelli2024new].
 
 # References
