### 2024-12-19 #7490 @ankushaggarwal
@jpsferreira Thanks! I have updated the paper as well as added contribution guidelines to the master branch. Please let me know if you any other feedback or comments.
