### 2024-12-25 #7490 @mpeirlinck
@ankushaggarwal

- To enhance the installation experience, it might be beneficial to recommend a specific Python version. I encountered package version conflicts with the default Python on anew conda env, but successfully installed the package using e.g. Python 3.9 with the following commands:

conda create -n pymecht python=3.9 ipykernel
conda activate pymecht
pip install pymecht

- The MCMC examples are very informative. To further support users, emphasizing the importance of selecting appropriate prior distributions and providing more detailed reasoning behind the default choices could be helpful. Additionally, a more explicit discussion on choosing burn-in sampling cutoff thresholds might guide users in making informed decisions.

- The paper section discussing advantages over finite element simulations is a bit trivial. To add more depth, it might be more beneficial to compare pymecht with existing analytical constitutive material fitting software, both open-source and commercial. Focusing on aspects like availability and ease of use could provide users with a clearer understanding of how pymecht stands out in the current software landscape.

- I would be excited to collaborate on integrating our universal material model architecture (https://doi.org/10.1007/s00366-024-02031-w) into the pymecht package. This integration could enable users to access billions of constitutive material models through a simple function, further enhancing the package's versatility.

- The ReadTheDocs documentation does a great job summarizing the Jupyter notebooks and examples. To further assist users who explore the code directly on git, consider numbering the example notebooks (e.g., 01_example.ipynb) to indicate a recommended sequence.

- To provide clarity on the collaborative effort behind the software, including a statement in the paper that outlines each author's contributions could be beneficial (from the git commits I only see three major contributors). This would help readers appreciate the diverse expertise that went into developing pymecht.
