### 2024-12-25 #7490 @siboles
@ankushaggarwal

### General Comments

- My overall impression of the pyMechT package is excellent.
- This package should prove valuable to researchers and developers in the continuum mechanics and numerical modelling fields. As the authors indicate, for simplified mechanical tests: uniaxial stress, biaxial extension, and tube extension/inflation, there is no need for inverse finite element analysis and a symbolic treatment is much less computationally expensive.
- The code is structured well and should allow for easy extension and contribution.
    - Adding type hints and/or utilizing a type validation library like pydantic would be a useful future improvement.
- The documentation is thorough and well-structured with excellent examples.
- The Bayesian approach using MCMC is of particular value experimentally. I recommend a sentence or two either in the paper and/or the documentation emphasizing the value of a confidence interval in the constitutive model fits. In particular, how confident one can be over a range of deformation.

### Installation

- Please update the instructions to include Python with the recommended version in the conda environment creation. As it stands the environment is empty without Python or pip.

    `conda create -n pymecht python=3.X`

- Using python setup.py develop failed for me due to the scipy version fetched not having an easy_install script. setup.py is deprecated now. Using instead,

    ```
    conda activate pymecht
    pip install -e .
    ```

    worked well for the development installation. I recommend updating to this installation method for installation from the repo.

### Minor Comments

- In the RTD example for [UniaxialExtension](https://pymecht.readthedocs.io/en/latest/Examples/Uniaxial_extension.html), the plot of Cauchy stress values using the 'strain' deformation type are suspiciously high compared to the stretch and delta options.
