+++ .github/workflows/ci-tests.yml
@@ -18,7 +18,7 @@ jobs:
       # will run even if one or more fail. 
       fail-fast: false
       matrix:
-        python-version: [3.8, 3.9, "3.10"]
+        python-version: [3.8, 3.9, "3.10", "3.11", "3.12", "3.13"]
         os: [ubuntu-latest, windows-latest, macos-latest]
     
     runs-on: ${{ matrix.os }}
