### 2025-01-05 ea09845
Take the Python version shield from the pyproject.toml file
