### 2025-01-13 d96c052
Added author contributions to the JOSS paper
