### 2025-01-16 fcb54a5
Reverting back to using pip install -e .
