### 2025-01-20 0825714
Added comparison to other constitutive model fitting software to JOSS paper, emphasising the advantages of pyMechT
