### 2025-01-20 #7490 @ankushaggarwal
@siboles @mpeirlinck Thank you so much for your review and excellent suggestions. We have made the following changes as per your comments:

- We have changed the installation instructions to include a python version [#40](https://github.com/ankushaggarwal/pymecht/pull/40)
- We have renamed the example notebooks to indicate an order [$41](https://github.com/ankushaggarwal/pymecht/pull/41)
- Fixed the range of strain in [UniaxialExtension](https://pymecht.readthedocs.io/en/latest/Examples/Uniaxial_extension.html) example [#39](https://github.com/ankushaggarwal/pymecht/pull/39)
- Added author contributions to the paper ([see commit](https://github.com/ankushaggarwal/pymecht/commit/d96c052f6cb38a2641f12f20553e0fbd2fbef8b2))
- Added comparison of `pyMechT` to other constitutive model fitting tools ([see commit](https://github.com/ankushaggarwal/pymecht/commit/0825714670c3345a1c8e4728c6b7989080b923e5))
- Added a section in the MCMC example explaining the importance of critical parameters (parameter ranges/variances and burn-in period) [#43](https://github.com/ankushaggarwal/pymecht/pull/43)

Please let us know if you have any other comments/suggestions/concerns. Thank you again.

@Kevin-Mattheus-Moerman
