### 2025-01-21 #7490 @mpeirlinck
@ankushaggarwal
Thanks for including our suggestions in the updated draft.

I just have one final note with respect to the comparison to other curve fitting tools: There are more tools than just hyperfit worth mentioning. E.g. both Dassault Systemes and Ansys have built-in material model fitting tools, Jorgen Bergstrom developed MCalibration and PolyUMod, and open-source alternatives include a.o. https://github.com/KnutAM/matmodfit https://github.com/adtzlr/hyperelastic
