### 2025-01-23 5657495
Updated JOSS paper to include other constitutive model fitting tools
