+++ joss-paper/paper.md
@@ -121,7 +121,8 @@ In principle, the problems that can be solved using `pyMechT` can also be solved
 
 - The reference zero-stress state of biological tissues can be unknown or ambiguous. Moreover, the biological tissues are heterogeneous, with multiple layers each of varying properties. These aspects are non-trivial to incorporate in a finite element simulation, due to the need for recreating the geometry and/or incompatability of the initial state. However, it is straightforward to simulate these in `pyMechT`.
 
-Additionally, there are alternative tools that can perform constitutive model fits. However, many of these are developed in-house and not aimed at general wider community development and usage. [Hyperfit](https://www.hyperfit.cz/home.php) is a commercial software for constitutive model fitting, with the advantage of having a graphical user interface, but is not free/open-source. Lastly, most of the existing tools do not incorporate Bayesian inference and layered structures, and these capabilities set `pyMechT` apart. 
+Overall, there are alternative tools that can perform constitutive model fitting. Commercial finite element software such as [Abaqus](https://www.3ds.com/products/simulia/abaqus) and [Ansys](https://www.ansys.com/) have in-built constitutive model fitting tools. For example, [PolyUMod](https://www.ansys.com/products/structures/polyumod) and [MCalibration](https://www.ansys.com/products/structures/mcalibration). [Hyperfit](https://www.hyperfit.cz/home.php) is a commercial software specifically for constitutive model fitting, with the advantage of having a graphical user interface. 
+However, these are not free/open-source. There are also open-source tools for constitutive model fitting, such as [matmodelfit](https://github.com/KnutAM/matmodfit/tree/master) and (hyperelastic)[https://github.com/adtzlr/hyperelastic]. However, these are not specifically focussed on tissues and lack the capability of handling layered samples or inflation-extension experiment on tubular structures, common in tissue mechanics. Lastly, most of the existing tools do not incorporate Bayesian inference. 
 
 # Uses in literature
 `pyMechT` has been used for Bayesian model selection based on extensive planar biaxial extension data [@AGGARWAL2023105657]. This work required rapid simulation of varied constitutive models, which was facilitated by `pyMechT`. Similarly, the Bayesian inference via Markov Chain Monte Carlo in `pyMechT` was used to infer the distribution of aortic biomechanical and geometrical properties based on in-vivo measurements (as likelihood) and ex-vivo biaxial extension data (as prior distribution) [@Aggarwal2025]. Moreover, data-driven model developed in @AGGARWAL2023115812 has been used in `pyMechT` via the `splineI1` and `splineI1I4` material models. 
