### 2025-01-23 c144f2b
Small fixes in the JOSS paper
