### 2025-01-23 5edb384
Small fix to the JOSS paper
