### 2025-01-23 #7490 @ankushaggarwal
@mpeirlinck Many thanks for pointing these out. I had heard of the fitting tools in Ansys/Abaqus, but I was not aware of the two github repositories you mentioned. I have extended the comparison section in the paper to include these now. Please let me know if you have any other comments.
