### 2025-02-06 #7490 @Kevin-Mattheus-Moerman
@ankushaggarwal with three positive reviews we can now proceed to process this work for acceptance in JOSS. I have just listed a set of checks/steps that we process at this point. Please can you review the steps assigned to you above and work on them? 👆
