### 2025-02-06 #7490 @ankushaggarwal
@Kevin-Mattheus-Moerman Great, thank you! I have completed the steps:

- Authors, affiliations, and ORCIDs are correct
- [v1.1.3](https://github.com/ankushaggarwal/pymecht/tree/v1.1.3) has been released
- Archived on Zenodo, doi: https://doi.org/10.5281/zenodo.14823424
- Title and author list is the same between JOSS paper and zenodo archive
- Same (MIT) license is used for the two

Please let me know if anything else is needed from my side.
