### 2025-02-07 #7490 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
