### 2025-02-11 #7490 @Kevin-Mattheus-Moerman
- [ ] check `incompatability`, this should probably read `incompatibility`.
- [ ] Words like `characterize` and `Conceptualization` show American English spelling in the paper. However `focussed` is British English. Consider using one type consistently.
