### 2025-02-12 4057603
Fixing two typos in the JOSS paper
