### 2025-02-12 #7490 @ankushaggarwal
@Kevin-Mattheus-Moerman Thanks for pointing these out. I have fixed them. I have also gone through the paper again to fix any other incosistent use of American/British English and typos (none were found).
