### 2025-02-14 33de371
Added badge for JOSS paper to README
